refactor: sectionController를 lineController 하위로 병합
