+++ src/main/java/subway/ui/LineController.java
@@ -3,11 +3,13 @@ package subway.ui;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.application.LineService;
+import subway.application.SectionService;
 import subway.dto.LineRequest;
 import subway.dto.LineResponse;
+import subway.dto.SectionSavingRequest;
+import subway.dto.StationRequest;
 
 import java.net.URI;
-import java.sql.SQLException;
 import java.util.List;
 
 @RestController
@@ -15,9 +17,12 @@ import java.util.List;
 public class LineController {
 
     private final LineService lineService;
+    private final SectionService sectionService;
 
-    public LineController(LineService lineService) {
+
+    public LineController(LineService lineService, SectionService sectionService) {
         this.lineService = lineService;
+        this.sectionService = sectionService;
     }
 
     @PostMapping
@@ -46,5 +51,18 @@ public class LineController {
     public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
         lineService.deleteLineById(id);
         return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/{lineId}/section")
+    public ResponseEntity<Void> initializeSection(@PathVariable long lineId, @RequestBody SectionSavingRequest sectionSavingRequest) {
+        long savedId = sectionService.insert(lineId, sectionSavingRequest.getPreviousStationName(),
+                sectionSavingRequest.getNextStationName(), sectionSavingRequest.getDistance(), sectionSavingRequest.isDown());
+        return ResponseEntity.created(URI.create(String.format("/lines/%d/%d", lineId, savedId))).build();
+    }
+
+    @DeleteMapping("/{lineId}/section")
+    public ResponseEntity<Void> deleteSection(@PathVariable long lineId, @RequestBody StationRequest stationRequest) {
+        sectionService.deleteStation(lineId, stationRequest.getName());
+        return ResponseEntity.ok().build();
     }
 }
